MINOR: buffer: add br_single() to check if a buffer ring has more than one buf
It's cheaper and cleaner than using br_count()==1 given that it just compares
two indexes, and that a ring having a single buffer is in a special case where
it is between empty and used up-to-1. In other words it's not congested.

(cherry picked from commit 9824f8c8908d67f6cedf2434e12a23f18a27eaf0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>