BUG/MINOR: quic: Missing STREAM frame data pointer updates
This patch follows this one which was not sufficient:
    "BUG/MINOR: quic: Missing STREAM frame length updates"
Indeed, it is not sufficient to update the ->len and ->offset member
of a STREAM frame to move it forward. The data pointer must also be updated.
This is not done by the STREAM frame builder.

Must be backported to 2.6 and 2.7.

(cherry picked from commit ca07979b978ed683b96d0ec99073c7cb972fd022)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit dcc827589a3a351af315d1f4cf0126dd2dc8a746)
[cf: ctx adjustment]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>