BUG/MEDIUM: mux-h2: do not try to free an unallocated h2s->sd
In h2s_close() we may dereference h2s->sd to get the sc, but this
function may be called on allocation error paths, so we must check
for this specific condition. Let's also update the comment to make
it explicitly permitted.

This needs to be backported to 2.6.

(cherry picked from commit bcdc6cc15b86dc89de4ca55244a2152386f1d1f1)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>