BUG/MINOR: stconn: fix sedesc memory leak on stream allocation failure
If we fail to allocate a new stream in sc_new_from_endp(), and the call
to sc_new() allocated the sedesc itself (which normally doesn't happen),
then it doesn't get released on the failure path. Let's explicitly
handle this case so that it's not overlooked and avoids some head
scratching sessions.

This may be backported to 2.6.

(cherry picked from commit 7a8ca0a0633e120c5b82de81f03c31cc39a17cb1)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>