MINOR: quic: Implement cubic state trace callback
This callback was left as not implemented. It should at least display
the algorithm state, the control congestion window the slow start threshold
and the time of the current recovery period. Should be helpful to debug.

Must be backported to 2.7.

(cherry picked from commit 01314b8b5367a08454a86262afe1311dbdb59706)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>