BUILD: da: extends CFLAGS to support API v3 from 3.1.7 and onwards.
Minor build update to still both support the v2 and v3 api from
the 3.1.7 release which supports a cache but would need a shift
in the HAProxy build not necessary at the moment.
In the second half of the year and for the next major HAProxy release
branch, v2 could be dropped altogether thus the next HAProxy 2.9
major release will contain more changes towards the v3 support
and reminder for the v2 EOL.

To be backported.

(cherry picked from commit cec3baa4fad82694edc255183872d2e02cfdaab3)
[cf: DEVICEATLAS_CFLAGS does not exist. OPTIONS_CFLAGS is change instead]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>